+++ test/State.ts
@@ -41,11 +41,11 @@ describe('State', () => {
   })
 
   it('eval', () => {
-    assert.deepStrictEqual(_.evalState(_.of<number, string>('a'), 0), 'a')
+    assert.deepStrictEqual(pipe(_.of<number, string>('a'), _.evalState(0)), 'a')
   })
 
   it('exec', () => {
-    assert.deepStrictEqual(_.execState(_.of<number, string>('a'), 0), 0)
+    assert.deepStrictEqual(pipe(_.of<number, string>('a'), _.execState(0)), 0)
   })
 
   it('put', () => {