+++ test/StateReaderTaskEither.ts
@@ -13,94 +13,96 @@ import * as T from '../src/Task'
 import * as TE from '../src/TaskEither'
 
 describe('StateReaderTaskEither', () => {
+  const defaultState: unknown = null
+
   describe('pipeables', () => {
     it('alt', async () => {
-      const e1 = await RTE.run(
-        _.evalState(
+      assert.deepStrictEqual(
+        await RTE.run(
           pipe(
-            _.right('a'),
-            _.alt(() => _.left(1))
+            _.right(1),
+            _.alt(() => _.left('a')),
+            _.evalState(defaultState)
           ),
           {}
         ),
-        {}
+        E.right(1)
       )
-      assert.deepStrictEqual(e1, E.right('a'))
-      const e2 = await RTE.run(
-        _.evalState(
+      assert.deepStrictEqual(
+        await RTE.run(
+          pipe(
+            _.left('a'),
+            _.alt(() => _.right(1)),
+            _.evalState(defaultState)
+          ),
+          {}
+        ),
+        E.right(1)
+      )
+      assert.deepStrictEqual(
+        await RTE.run(
           pipe(
             _.left(1),
-            _.alt(() => _.right('b'))
+            _.alt(() => _.left(2)),
+            _.evalState(defaultState)
           ),
           {}
         ),
-        {}
+        E.left(2)
       )
-      assert.deepStrictEqual(e2, E.right('b'))
-      const e3 = await RTE.run(
-        _.evalState(
-          pipe(
-            _.left(1),
-            _.alt(() => _.left(2))
-          ),
-          {}
-        ),
-        {}
-      )
-      assert.deepStrictEqual(e3, E.left(2))
     })
 
     it('map', async () => {
       const len = (s: string): number => s.length
-      const e = await RTE.run(_.evalState(pipe(_.right('aaa'), _.map(len)), {}), {})
+      const e = await RTE.run(pipe(_.right('aaa'), _.map(len), _.evalState(defaultState)), {})
       assert.deepStrictEqual(e, E.right(3))
     })
 
     it('ap', async () => {
       const len = (s: string): number => s.length
-      const e = await RTE.run(_.evalState(pipe(_.right(len), _.ap(_.right('aaa'))), {}), {})
+      const e = await RTE.run(pipe(_.right(len), _.ap(_.right('aaa')), _.evalState(defaultState)), {})
       assert.deepStrictEqual(e, E.right(3))
     })
 
     it('apFirst', async () => {
-      const e = await RTE.run(_.evalState(pipe(_.right('a'), _.apFirst(_.right('b'))), {}), {})
+      const e = await RTE.run(pipe(_.right('a'), _.apFirst(_.right('b')), _.evalState(defaultState)), {})
       assert.deepStrictEqual(e, E.right('a'))
     })
 
     it('apSecond', async () => {
-      const e = await RTE.run(_.evalState(pipe(_.right('a'), _.apSecond(_.right('b'))), {}), {})
+      const e = await RTE.run(pipe(_.right('a'), _.apSecond(_.right('b')), _.evalState(defaultState)), {})
       assert.deepStrictEqual(e, E.right('b'))
     })
 
     it('chain', async () => {
       const f = (s: string) => (s.length > 2 ? _.right(s.length) : _.right(0))
-      const e = await RTE.run(_.evalState(pipe(_.right('aaa'), _.chain(f)), {}), {})
+      const e = await RTE.run(pipe(_.right('aaa'), _.chain(f), _.evalState(defaultState)), {})
       assert.deepStrictEqual(e, E.right(3))
     })
 
     it('chainFirst', async () => {
       const f = (s: string) => (s.length > 2 ? _.right(s.length) : _.right(0))
-      const e = await RTE.run(_.evalState(pipe(_.right('aaa'), _.chainFirst(f)), {}), {})
+      const e = await RTE.run(pipe(_.right('aaa'), _.chainFirst(f), _.evalState(defaultState)), {})
       assert.deepStrictEqual(e, E.right('aaa'))
     })
 
     it('chainFirst', async () => {
-      const e = await RTE.run(_.evalState(pipe(_.right(_.right('a')), _.flatten), {}), {})
+      const e = await RTE.run(pipe(_.right(_.right('a')), _.flatten, _.evalState(defaultState)), {})
       assert.deepStrictEqual(e, E.right('a'))
     })
 
     it('bimap', async () => {
       const gt2 = (n: number): boolean => n > 2
       const len = (s: string): number => s.length
-      const e1 = await RTE.run(_.evalState(pipe(_.right('aaa'), _.bimap(gt2, len)), {}), {})
+      const e1 = await RTE.run(pipe(_.right('aaa'), _.bimap(gt2, len), _.evalState(defaultState)), {})
       assert.deepStrictEqual(e1, E.right(3))
-      const e2 = await RTE.run(_.evalState(pipe(_.left(3), _.bimap(gt2, len)), {}), {})
+      const e2 = await RTE.run(pipe(_.left(3), _.bimap(gt2, len), _.evalState(defaultState)), {})
       assert.deepStrictEqual(e2, E.left(true))
     })
 
     it('mapLeft', async () => {
       const gt2 = (n: number): boolean => n > 2
-      const e = await RTE.run(_.evalState(pipe(_.left(3), _.mapLeft(gt2)), {}), {})
+      const e = await RTE.run(pipe(_.left(3), _.mapLeft(gt2), _.evalState(defaultState)), {})
       assert.deepStrictEqual(e, E.left(true))
     })
 
@@ -111,83 +113,80 @@ describe('StateReaderTaskEither', () => {
       const refinement = (u: string | number): u is number => typeof u === 'number'
       const isNumber = _.fromPredicate(refinement, (u) => `Invalid number ${String(u)}`)
 
-      const e1 = await RTE.run(_.evalState(gt2(3), {}), {})
-      const e2 = await RTE.run(_.evalState(gt2(1), {}), {})
-      const e3 = await RTE.run(_.evalState(isNumber(4), {}), {})
+      const e1 = await RTE.run(_.evalState(defaultState)(gt2(3)), {})
+      const e2 = await RTE.run(_.evalState(defaultState)(gt2(1)), {})
+      const e3 = await RTE.run(_.evalState(defaultState)(isNumber(4)), {})
       assert.deepStrictEqual(e1, E.right(3))
       assert.deepStrictEqual(e2, E.left('Invalid number 1'))
       assert.deepStrictEqual(e3, E.right(4))
     })
 
     it('filterOrElse', async () => {
-      const e1 = await RTE.run(
-        _.evalState(
+      assert.deepStrictEqual(
+        await RTE.run(
           pipe(
             _.right(12),
             _.filterOrElse(
               (n) => n > 10,
               () => 'a'
-            )
+            ),
+            _.evalState(defaultState)
           ),
           {}
         ),
-        {}
+        E.right(12)
       )
-      assert.deepStrictEqual(e1, E.right(12))
 
-      const e2 = await RTE.run(
-        _.evalState(
+      assert.deepStrictEqual(
+        await RTE.run(
           pipe(
             _.right(8),
             _.filterOrElse(
               (n) => n > 10,
               () => 'a'
-            )
+            ),
+            _.evalState(defaultState)
           ),
           {}
         ),
-        {}
+        E.left('a')
       )
-      assert.deepStrictEqual(e2, E.left('a'))
     })
   })
 
   it('run', async () => {
-    const ma = _.right('aaa')
-    const e = await _.run(ma, {}, {})
-    assert.deepStrictEqual(e, E.right(['aaa', {}]))
+    assert.deepStrictEqual(await _.run(_.right('a'), {}, {}), E.right(['a', {}]))
   })
 
   it('ap (seq)', async () => {
     const len = (s: string): number => s.length
     const mab = _.right(len)
-    const ma = _.right('aaa')
-    const e = await RTE.run(_.evalState(pipe(mab, _.monadReaderTaskEitherSeq.ap(ma)), {}), {})
-    assert.deepStrictEqual(e, E.right(3))
+    const ma = _.right('a')
+    assert.deepStrictEqual(
+      await RTE.run(pipe(mab, _.monadReaderTaskEitherSeq.ap(ma), _.evalState(defaultState)), {}),
+      E.right(1)
+    )
   })
 
   it('execState', async () => {
-    const ma = _.right('aaa')
-    const s = {}
-    const e = await RTE.run(_.execState(ma, s), {})
-    assert.deepStrictEqual(e, E.right(s))
+    assert.deepStrictEqual(await RTE.run(_.execState(defaultState)(_.right(1)), {}), E.right(null))
   })
 
   it('rightState', async () => {
-    const state: State<{}, number> = (s) => [1, s]
-    const e = await RTE.run(_.evalState(_.rightState(state), {}), {})
+    const state: State<unknown, number> = (s) => [1, s]
+    const e = await RTE.run(_.evalState(defaultState)(_.rightState(state)), {})
     assert.deepStrictEqual(e, E.right(1))
   })
 
   it('leftState', async () => {
-    const state: State<{}, number> = (s) => [1, s]
-    const e = await RTE.run(_.evalState(_.leftState(state), {}), {})
+    const state: State<unknown, number> = (s) => [1, s]
+    const e = await RTE.run(_.evalState(defaultState)(_.leftState(state)), {})
     assert.deepStrictEqual(e, E.left(1))
   })
 
   it('fromReaderTaskEither', async () => {
-    const rte: RTE.ReaderTaskEither<{}, string, number> = RTE.right(1)
-    const e = await RTE.run(_.evalState(_.fromReaderTaskEither(rte), {}), {})
+    const rte: RTE.ReaderTaskEither<unknown, string, number> = RTE.right(1)
+    const e = await RTE.run(_.evalState(defaultState)(_.fromReaderTaskEither(rte)), {})
     assert.deepStrictEqual(e, E.right(1))
   })